requirements.txt
torch
torchvision
pillow
numpy
streamlit
tqdm
nltk
rouge-score
matplotlib
scikit-learn
pandas